backend/requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
django-phonenumber-field==7.0.0
djangorestframework==3.15.1
environs[django]==9.3.2
geopy==2.4.1
gunicorn==22.0.0
phonenumbers==8.13.37
pillow==10.3.0
psycopg2==2.9.9
requests==2.32.2
rollbar==1.0.0
